+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/ReverseGeocodingSearchTest.kt
@@ -6,8 +6,6 @@ import com.mapbox.search.base.core.CoreApiType
 import com.mapbox.search.base.core.CoreReverseGeoOptions
 import com.mapbox.search.base.core.CoreSearchCallback
 import com.mapbox.search.base.core.CoreSearchEngineInterface
-import com.mapbox.search.base.logger.reinitializeLogImpl
-import com.mapbox.search.base.logger.resetLogImpl
 import com.mapbox.search.base.result.BaseGeocodingCompatSearchSuggestion
 import com.mapbox.search.base.result.BaseSearchSuggestion
 import com.mapbox.search.base.result.SearchRequestContext
@@ -16,7 +14,6 @@ import com.mapbox.search.base.result.mapToBase
 import com.mapbox.search.base.task.AsyncOperationTaskImpl
 import com.mapbox.search.common.SearchCancellationException
 import com.mapbox.search.common.SearchRequestException
-import com.mapbox.search.common.tests.TestConstants.ASSERTIONS_KT_CLASS_NAME
 import com.mapbox.search.common.tests.TestExecutor
 import com.mapbox.search.common.tests.TestThreadExecutorService
 import com.mapbox.search.common.tests.createCoreSearchAddress
@@ -32,12 +29,8 @@ import com.mapbox.search.tests_support.createTestServerSearchResult
 import com.mapbox.test.dsl.TestCase
 import io.mockk.every
 import io.mockk.mockk
-import io.mockk.mockkStatic
 import io.mockk.slot
 import io.mockk.spyk
-import io.mockk.unmockkStatic
-import org.junit.jupiter.api.AfterAll
-import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.TestFactory
 import java.util.concurrent.Executor
@@ -364,21 +357,5 @@ internal class ReverseGeocodingSearchTest {
             TEST_REQUEST_OPTIONS.mapToCore(),
             TEST_RESPONSE_UUID
         )
-
-        @Suppress("JVM_STATIC_IN_PRIVATE_COMPANION")
-        @BeforeAll
-        @JvmStatic
-        fun setUpAll() {
-            resetLogImpl()
-            mockkStatic(ASSERTIONS_KT_CLASS_NAME)
-        }
-
-        @Suppress("JVM_STATIC_IN_PRIVATE_COMPANION")
-        @AfterAll
-        @JvmStatic
-        fun tearDownAll() {
-            reinitializeLogImpl()
-            unmockkStatic(ASSERTIONS_KT_CLASS_NAME)
-        }
     }
 }